# backend/swagger.yaml
openapi: 3.0.0
info:
  title: swagger
  version: 1.0.0
servers:
  - url: https://yeti-jobs.onrender.com/api/v1/
paths:
...
  /admin/search/users:
    get:
      summary: Search Users
      operationId: searchUsers
      description: ""
      tags:
        - Admin
      responses:
        "200":
          description: ""
      parameters:
        - name: q
          in: query
          description: ""
          required: true
          schema:
            type: string
          example: g
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT